pyproject.toml
[tool.poetry]
name = "tortoise-pagination"
version = "1.2.11"
description = "Pagination for Tortoise-ORM on FastAPI"
authors = ["Sebastien Nicolet <snicolet95@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tortoise_pagination"}]

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.21.0"
fastapi = ">=0.104.1"
pydantic = ">2.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
flake8 = "^6.0.0"
black = "^23.11.0"
pytest = "^8.3.2"
testcontainers = "^4.8.1"
mock = "^5.1.0"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.24.0"
anyio = "^4.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[rools.ruff]
line_length = 100

[tool.black]
line_length = 100
